requirements.txt
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
dj-pagination==2.3.0
Django==1.10
django-comments-xtd==2.0.6
django-contrib-comments==1.8.0
django-modelcluster==3.1
django-taggit==0.22.1
django-treebeard==4.1.2
djangorestframework==3.6.4
docutils==0.14
elasticsearch==5.4.0
html5lib==0.999999999
idna==2.6
Markdown==2.6.9
olefile==0.44
Pillow==4.2.1
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
six==1.10.0
Unidecode==0.4.21
urllib3==1.22
wagtail==1.12.1
webencodings==0.5.1
Willow==0.4
